flask-app/requirements.txt
flask==3.0.3
requests==2.32.3
python-jose[cryptography]==3.3.0
authlib==1.3.1
